pyproject.toml
[project]
name = "agent-native-speech"
version = "0.1.0"
description = "MCP server that lets AI agents call you on Discord voice channels"
requires-python = ">=3.10"
dependencies = [
    "discord.py[voice]>=2.5",
    "discord-ext-voice-recv>=0.5.1",
    "faster-whisper>=1.1.0",
    "torch>=2.0",
    "librosa>=0.10",
    "soundfile>=0.12",
    "numpy>=1.24",
    "anthropic>=0.40",
    "mcp>=1.0",
    "python-dotenv>=1.0",
    "starlette>=0.36",
    "uvicorn>=0.27",
    "thefuzz>=0.22",
    "python-Levenshtein>=0.25",
    "httpx>=0.27",
]

[project.optional-dependencies]
tts = [
    "qwen-tts>=0.1",
    "flash-attn>=2.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
voice-agent = "server.init:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["server*"]
